gateway/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.6.1
slowapi==0.1.9
